refactor: divide actions from reducer files
