+++ src/pages/Articles/ArticleList.tsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useInfiniteScroll } from 'src/hooks'
 import { PropsTypes } from 'src/types'
 import { apiUtils } from 'src/utils'
-import { getArticles } from 'src/services/articles/reducer'
+import { getArticles } from 'src/services/articles/actions'
 import { articlesSelector } from 'src/services/articles/selectors'
 import ArticleListItem from './ArticleListItem'
 import ArticlesLoading from 'src/pages/Articles/ArticlesLoading'
