+++ src/pages/Home/EmailFormSection/EmilFormModal.tsx
@@ -4,7 +4,7 @@ import { Modal, Input, Button } from 'antd'
 import styled from 'styled-components'
 import { validators } from 'src/utils'
 import { keyCodes } from 'src/constants'
-import { sendEmailSubscribe } from 'src/services/subscribe/reducer'
+import { sendEmailSubscribe } from 'src/services/subscribe/actions'
 
 interface Props {
   closeModal: () => void
