+++ src/services/subscribe/saga.ts
@@ -8,7 +8,7 @@ import {
   sendEmailSubscribeFailure,
   startSubscribeLoading,
   endSubscribeLoading,
-} from 'src/services/subscribe/reducer'
+} from 'src/services/subscribe/actions'
 
 // TODO: 사가 함수에 액션 파라미터를 넘길 때 타입이 먹히지 않는 이슈 아직 유효 한 지 확인하기
 // https://github.com/redux-saga/redux-saga/issues/1188
